pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auresnet-dz"
version = "0.1.0"
description = "CCLM/CORDEX to ERA5 post-processing for the Aures (NE Algeria) region"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "AuresNet-DZ Team" }]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "xarray>=2024.3.0",
  "cdsapi>=0.7.3",
  "dask[distributed]>=2024.3.0",
  "netCDF4>=1.6.5",
  "h5netcdf>=1.3.0",
  "zarr>=2.18.2",
  "cf_xarray>=0.9.0",
  "scipy>=1.13",
  "xesmf>=0.8.5",
  "torch>=2.3",
  "torchvision>=0.18",
  "pytorch-lightning>=2.2",
  "segmentation-models-pytorch>=0.3.3",
  "einops>=0.8.0",
  "pyyaml>=6.0.1",
  "hydra-core>=1.3.2",
  "rich>=13.7",
  "tqdm>=4.66",
  "tensorboard>=2.16",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
